test: move tests for setAllowedCalls/ERC725YDataKeys in their own file in LSP20
